+++ assembly/src/tests.rs
@@ -7,9 +7,9 @@ use miden_assembly_syntax::{
     diagnostics::WrapErr,
 };
 use miden_core::{
-    Operation, Program, Word, assert_matches,
+    EventId, Operation, Program, Word, assert_matches,
     mast::{MastNode, MastNodeExt, MastNodeId},
-    utils::{Deserializable, Serializable, string_to_event_id},
+    utils::{Deserializable, Serializable},
 };
 use miden_mast_package::{MastArtifact, MastForest, Package, PackageExport, PackageManifest};
 use proptest::{
@@ -1246,7 +1246,7 @@ fn const_word_from_string() -> TestResult {
 fn const_event_from_string() -> TestResult {
     let context = TestContext::default();
     let sample_event_name = "miden::test::constant";
-    let expected_felt = string_to_event_id(sample_event_name);
+    let expected_felt = EventId::from_name(sample_event_name);
 
     let source1 = source_file!(
         &context,