+++ core/src/sys_events.rs
@@ -7,7 +7,7 @@ use core::fmt;
 // between 0 and 2^32.
 pub use constants::*;
 
-use crate::Felt;
+use crate::EventId;
 
 #[rustfmt::skip]
 mod constants {
@@ -339,8 +339,8 @@ impl SystemEvent {
 
     /// Returns a system event corresponding to the specified event ID, or `None` if the event
     /// ID is not recognized.
-    pub fn from_event_id(event_id: Felt) -> Option<Self> {
-        let event_id: u32 = event_id.as_int().try_into().ok()?;
+    pub fn from_event_id(event_id: EventId) -> Option<Self> {
+        let event_id: u32 = event_id.as_felt().as_int().try_into().ok()?;
         match event_id {
             EVENT_MERKLE_NODE_MERGE => Some(SystemEvent::MerkleNodeMerge),
             EVENT_MERKLE_NODE_TO_STACK => Some(SystemEvent::MerkleNodeToStack),