+++ processor/src/host/default.rs
@@ -1,6 +1,6 @@
 use alloc::{sync::Arc, vec::Vec};
 
-use miden_core::{DebugOptions, Felt, Word, mast::MastForest};
+use miden_core::{DebugOptions, EventId, Felt, Word, mast::MastForest};
 use miden_debug_types::{
     DefaultSourceManager, Location, SourceFile, SourceManager, SourceManagerSync, SourceSpan,
 };
@@ -80,7 +80,7 @@ where
     /// `fn(&mut ProcessState) -> Result<(), EventHandler>`
     pub fn register_handler(
         &mut self,
-        id: Felt,
+        id: EventId,
         handler: Arc<dyn EventHandler>,
     ) -> Result<(), ExecutionError> {
         self.event_handlers.register(id, handler)
@@ -88,13 +88,13 @@ where
 
     /// Un-registers a handler with the given id, returning a flag indicating whether a handler
     /// was previously registered with this id.
-    pub fn unregister_handler(&mut self, id: Felt) -> bool {
+    pub fn unregister_handler(&mut self, id: EventId) -> bool {
         self.event_handlers.unregister(id)
     }
 
     /// Replaces a handler with the given id, returning a flag indicating whether a handler
     /// was previously registered with this id.
-    pub fn replace_handler(&mut self, id: Felt, handler: Arc<dyn EventHandler>) -> bool {
+    pub fn replace_handler(&mut self, id: EventId, handler: Arc<dyn EventHandler>) -> bool {
         let existed = self.event_handlers.unregister(id);
         self.register_handler(id, handler).unwrap();
         existed
@@ -155,7 +155,7 @@ where
     }
 
     fn on_event(&mut self, process: &ProcessState) -> Result<Vec<AdviceMutation>, EventError> {
-        let event_id = process.get_stack_item(0);
+        let event_id = EventId::from_felt(process.get_stack_item(0));
         if let Some(mutations) = self.event_handlers.handle_event(event_id, process)? {
             // the event was handled by the registered event handlers; just return
             return Ok(mutations);
@@ -199,7 +199,7 @@ pub struct HostLibrary {
     /// A `MastForest` with procedures exposed by this library.
     pub mast_forest: Arc<MastForest>,
     /// List of handlers along with an event id to call them with `emit`.
-    pub handlers: Vec<(Felt, Arc<dyn EventHandler>)>,
+    pub handlers: Vec<(EventId, Arc<dyn EventHandler>)>,
 }
 
 impl From<Arc<MastForest>> for HostLibrary {